refactor: modularize analyzer into focused single-responsibility modules
Split monolithic analyzer.go (1,777 lines) into 12 focused modules:
- ast_walker.go (161 lines) - AST traversal utilities
- cache.go (94 lines) - template caching
- config.go (32 lines) - configuration types
- context_checks.go (94 lines) - context usage validation
- coordinator.go (53 lines) - check orchestration
- diagnostics.go (60 lines) - diagnostic constants
- error_checks.go (290 lines) - error logging validation
- property_checks.go (228 lines) - property naming/capturing
- quickfix.go (407 lines) - suggested fix generation
- template_checks.go (258 lines) - template argument validation
- validation.go (105 lines) - helper validation functions
- analyzer.go reduced to 128 lines
